6-8视频

+++ app/Http/Controllers/PostController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use \App\Post;
+use \App\User;
 
 class PostController extends Controller
 {
@@ -24,7 +25,9 @@ class PostController extends Controller
     //文章详情页
     public function show(Post $post)
     {
-        return view('post/show', compact('post'));
+        $userId = \Auth::id();
+        $user = User::find($userId);
+        return view('post/show', compact('post', 'user'));
     }
 
     //创建文章页面
@@ -40,7 +43,9 @@ class PostController extends Controller
             'title' => 'required | string | max:100 | min:5',
             'content' => 'required | string | min:10',
         ]);
-        $return = Post::create(request(['title', 'content']));
+        $userId = \Auth::id();
+        $params = array_merge(request(['title', 'content']), ['user_id' =>$userId]);
+        $return = Post::create($params);
         return redirect('/posts');
     }
 
@@ -53,7 +58,8 @@ class PostController extends Controller
     //编辑文章
     public function update(Post $post)
     {
-        //Todo: 用户权限认证
+        //用户权限认证
+        $this->authorize('update', $post);
         $this->validate(request(), [
             'title' => 'required | string | max:100 | min:5',
             'content' => 'required | string | min:10',
@@ -67,7 +73,8 @@ class PostController extends Controller
     //删除文章
     public function delete(Post $post)
     {
-        //Todo: 用户权限认证
+        //用户权限认证
+        $this->authorize('delete', $post);
         $post->delete();
         return redirect('/posts');
     }
